Send buyer submission emails outside database transaction
This changes the buyer submission service object so that sending the manager approval email happens after the database transaction has been completed.

When this has not been the case, the manager approval token is not included in the email body, as it's not able to be picked up by the mailer object, running in the worker queue.